Set up for lecture 12
