+++ controller-web.ts
@@ -1,0 +1,7 @@
+import {Model} from './model';
+import {View} from './view';
+
+//Fill me in!!
+export class Controller {
+
+}
